Change types to fix compile warnings

+++ src/Averager.cpp
@@ -15,9 +15,9 @@ void SerialWeightedMultiVectorAverager::calculateSum(const Matrix* const data, M
                                                      const std::vector<int>* const dataAssignments,
                                                      const std::vector<value_t>* const weights)
 {
-    for (int i = 0; i < data->getNumData(); i++)
+    for (int_fast32_t i = 0; i < data->getNumData(); i++)
     {
-        for (int j = 0; j < data->getNumFeatures(); j++)
+        for (int_fast32_t j = 0; j < data->getNumFeatures(); j++)
         {
             avgContainer->at(dataAssignments->at(i), j) += weights->at(i) * data->at(i, j);
         }
@@ -27,9 +27,9 @@ void SerialWeightedMultiVectorAverager::calculateSum(const Matrix* const data, M
 void SerialWeightedMultiVectorAverager::normalizeSum(Matrix* const avgContainer,
                                                      const std::vector<value_t>* const weightSums)
 {
-    for (int i = 0; i < avgContainer->getNumData(); i++)
+    for (int_fast32_t i = 0; i < avgContainer->getNumData(); i++)
     {
-        for (int j = 0; j < avgContainer->getNumFeatures(); j++)
+        for (int_fast32_t j = 0; j < avgContainer->getNumFeatures(); j++)
         {
             avgContainer->at(i, j) /= weightSums->at(i);
         }
@@ -40,9 +40,9 @@ void OMPWeightedMultiVectorAverager::normalizeSum(Matrix* const avgContainer,
                                                   const std::vector<value_t>* const weightSums)
 {
 #pragma omp parallel for shared(avgContainer, weightSums), schedule(static), collapse(2)
-    for (int i = 0; i < avgContainer->getNumData(); i++)
+    for (int_fast32_t i = 0; i < avgContainer->getNumData(); i++)
     {
-        for (int j = 0; j < avgContainer->getNumFeatures(); j++)
+        for (int_fast32_t j = 0; j < avgContainer->getNumFeatures(); j++)
         {
             avgContainer->at(i, j) /= weightSums->at(i);
         }
@@ -57,9 +57,9 @@ void AbstractAverager::calculateAverage(const Matrix* const data, std::vector<va
 
 void SerialVectorAverager::calculateSum(const Matrix* const data, std::vector<value_t>* const avgContainer)
 {
-    for (int i = 0; i < data->getNumData(); i++)
+    for (int_fast32_t i = 0; i < data->getNumData(); i++)
     {
-        for (int j = 0; j < data->getNumFeatures(); j++)
+        for (int_fast32_t j = 0; j < data->getNumFeatures(); j++)
         {
             avgContainer->at(j) += data->at(i, j);
         }
@@ -68,7 +68,7 @@ void SerialVectorAverager::calculateSum(const Matrix* const data, std::vector<va
 
 void SerialVectorAverager::normalizeSum(std::vector<value_t>* const avgContainer, const int numData)
 {
-    for (int i = 0; i < avgContainer->size(); i++)
+    for (size_t i = 0; i < avgContainer->size(); i++)
     {
         avgContainer->at(i) /= numData;
     }
@@ -80,9 +80,9 @@ void OMPVectorAverager::calculateSum(const Matrix* const data, std::vector<value
     auto copyContainerData = copyContainer.data();
 
 #pragma omp parallel for shared(data, avgContainer), schedule(static), collapse(2), reduction(+ : copyContainerData[:data->getNumFeatures()])
-    for (int i = 0; i < data->getNumData(); i++)
+    for (int_fast32_t i = 0; i < data->getNumData(); i++)
     {
-        for (int j = 0; j < data->getNumFeatures(); j++)
+        for (int_fast32_t j = 0; j < data->getNumFeatures(); j++)
         {
             copyContainer[j] += data->at(i, j);
         }
@@ -94,7 +94,7 @@ void OMPVectorAverager::calculateSum(const Matrix* const data, std::vector<value
 void OMPVectorAverager::normalizeSum(std::vector<value_t>* const avgContainer, const int numData)
 {
 #pragma omp parallel for shared(avgContainer, numData), schedule(static)
-    for (int i = 0; i < avgContainer->size(); i++)
+    for (size_t i = 0; i < avgContainer->size(); i++)
     {
         avgContainer->at(i) /= numData;
     }
